updated to use 64-bit version

+++ cicada/grammar_static.cpp
@@ -410,7 +410,7 @@ namespace cicada
     
     const rule_pair_set_type& read_rule_set(size_type node) const
     {
-      typedef utils::hashmurmur3<uint32_t> hasher_type;
+      typedef utils::hashmurmur3<size_t> hasher_type;
 
       FeatureVectorCODEC   feature_codec;
       AttributeVectorCODEC attribute_codec;
@@ -542,7 +542,7 @@ namespace cicada
 				     const cache_phrase_set_type& cache_phrases,
 				     const phrase_db_type& phrase_db) const
     {
-      typedef utils::hashmurmur3<uint32_t> hasher_type;
+      typedef utils::hashmurmur3<size_t> hasher_type;
       
       const size_type cache_pos = hasher_type()(pos, lhs.id()) & (cache_phrases.size() - 1);
